backend/requirements.txt
fastapi==0.135.3
uvicorn==0.43.0
motor==3.7.1
redis==5.0.1
upstash-redis==1.4.0
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.2
python-socketio[asgi]==5.11.1
slowapi==0.1.9
httpx==0.28.1
PyJWT[crypto]==2.12.1
